chore: fix bug

+++ chain/connector/src/connector/mod.rs
@@ -38,7 +38,7 @@ type EventsChannel = (
 
 const MIN_CONNECTION_TIMEOUT: Duration = Duration::from_millis(100);
 const MIN_TX_CONFIRM_TIMEOUT: Duration = Duration::from_secs(1);
-const TX_TIMEOUT_MULTIPLIER: u64 = 2;
+const TX_TIMEOUT_MULTIPLIER: u32 = 2;
 const DEFAULT_SYNC_TOLERANCE_PCT: usize = 90;
 
 /// Configuration of the [`HoprBlockchainConnector`].
